[esbmc] split command-line option handling into parseoptions/
Moves get_command_line_options, read_time_spec and read_mem_spec,
together with the file-local statics esbmc_version_string,
format_target, timeout_handler and segfault_handler (and the
BT_BUF_SIZE macro plus the buildidstring_buf externs they use), into
src/esbmc/parseoptions/command_line_options.cpp.

Verbatim move. --help and --version output is byte-identical to master,
the SIGALRM timeout path still fires, and the esbmc suite reproduces the
same 11 pre-existing failures as master.
