+++ src/esbmc/CMakeLists.txt
@@ -44,6 +44,7 @@ add_custom_command(OUTPUT ${CMAKE_CURRENT_BINARY_DIR}/buildidobj.c
 
 add_executable(esbmc main.cpp esbmc_parseoptions.cpp
     parseoptions/bmc_strategy.cpp
+    parseoptions/command_line_options.cpp
     parseoptions/function_contracts.cpp
     parseoptions/goto_program.cpp
     parseoptions/k_induction.cpp
